test(app): move the backoff test beside failureLog
The test never touched the loop, the harness or the logger: it drives
failureLog directly and asserts the run lengths it reports. Sitting in
app_test.go it read as a test of the poll loop, which is the one thing
it does not exercise.
